* Change a lot of positions to be stored as integers instead of floats * Increase research costs for many things * Start working on save and load features * Increase granularity of unitsInBigSquares and make it use vectors instead of sets * Minimalistic XML library created, for use when saving and loading save files
git-svn-id: https://bloopz.zapto.org/svn/nightfall/projekt/trunk@949 d57568f3-fe0c-0410-b608-e86116ac1d00

+++ src/aibase.h
@@ -87,11 +87,11 @@ namespace Game
 		void PerformAI(Dimension::Unit* unit);     // Execute AI stuff for a unit
 		void PerformAI(Dimension::Player* player); // Execute player AI (things that have nothing to do with the units, that is)
 		void PerformAIFrame(); // Perform a full AI frame, if possible
-		void CommandUnit(Dimension::Unit* unit, float x, float y, UnitAction action, void* argument, bool queue, bool insert);
+		void CommandUnit(Dimension::Unit* unit, int x, int y, UnitAction action, void* argument, bool queue, bool insert);
 		// command/instruct a unit to do something at (x, y)
 		void CommandUnit(Dimension::Unit* unit, Dimension::Unit* destination, UnitAction action, void* argument, bool queue, bool insert); 
 		// command/instruct the unit to do something with another unit
-		void CommandUnits(vector<Dimension::Unit*> units, float x, float y, UnitAction action, void* argument, bool queue, bool insert);
+		void CommandUnits(vector<Dimension::Unit*> units, int x, int y, UnitAction action, void* argument, bool queue, bool insert);
 		// command/instruct several units to do something at (x, y)
 		void CommandUnits(vector<Dimension::Unit*> units, Dimension::Unit* destination, UnitAction action, void* argument, bool queue, bool insert); 
 		// command/instruct several units to do something with another unit
@@ -102,7 +102,7 @@ namespace Game
 		void CompleteAction(Dimension::Unit* pUnit);
 		void CancelAllActions(Dimension::Unit* pUnit);
 		void ScheduleNextAction(Dimension::Unit* pUnit);
-		void ApplyAction(Dimension::Unit* pUnit, UnitAction action, float goal_x, float goal_y, Dimension::Unit* target, void* arg);
+		void ApplyAction(Dimension::Unit* pUnit, UnitAction action, int goal_x, int goal_y, Dimension::Unit* target, void* arg);
 		
 		void SendUnitEventToLua_BuildComplete(Dimension::Unit* pUnit);
 		void SendUnitEventToLua_BuildCancelled(Dimension::Unit* pUnit);
